requirements.txt
fastapi
uvicorn
pymysql
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
httpx
python-multipart
